\name{withExclusions-package}
\alias{withExclusions-package}
\alias{withExclusions}
\docType{package}
\title{
\packageTitle{withExclusions}
}
\description{
\packageDescription{withExclusions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{withExclusions}
\packageIndices{withExclusions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{withExclusions}

Maintainer: \packageMaintainer{withExclusions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
